feat(review-pacing): lay the tab out in two columns with a preset header
Replaces the single narrow column and its height-pushing advanced accordion.
A preset line heads the tab, then General and Daily Limits sit side by side
above 46rem of container width. Advanced crossfades into the right column in
place via crossfade-resize, so the tab no longer scrolls above the sheet.

Extracts general-section and limits-section out of the old pacing-section,
and unifies the tab's copy on 'Review' / 'Scheduling'.
