+++ src/locales/en-us.json
@@ -547,7 +547,7 @@
   "deck.settings-modal.aside.date-fallback": "—",
   "deck.settings-modal.tab.design": "Appearance",
   "deck.settings-modal.tab.details": "Details",
-  "deck.settings-modal.tab.review-pacing": "Review Settings",
+  "deck.settings-modal.tab.review-pacing": "Review",
   "deck.settings-modal.tab.review-history": "Review History",
   "deck.settings-modal.tab.danger-zone": "Danger Zone",
   "deck.settings-modal.review-history.coming-soon": "Coming very soon",
@@ -584,18 +584,22 @@
   "deck.settings-modal.review-pacing.starting-side.back": "Back",
   "deck.settings-modal.review-pacing.starting-side.front": "Front",
   "deck.settings-modal.review-pacing.starting-side.random": "Random",
-  "deck.settings-modal.review-pacing.retry-failed": "Retry Failed Cards",
   "deck.settings-modal.review-pacing.max-reviews-per-day": "Max Reviews",
   "deck.settings-modal.review-pacing.max-reviews-tooltip": "Caps how many due cards you'll review today. Extra cards roll over to tomorrow. Set to 0 for no limit.",
   "deck.settings-modal.review-pacing.max-new-per-day": "Max New Cards",
   "deck.settings-modal.review-pacing.max-new-tooltip": "Caps how many new cards you'll see today. The rest wait until tomorrow. Set to 0 for no limit.",
-  "deck.settings-modal.review-pacing.section.cards-heading": "General",
-  "deck.settings-modal.review-pacing.section-heading": "Pacing",
-  "deck.settings-modal.review-pacing.section-description": "Choose a preset to control how this deck schedules reviews, or fine-tune it below.",
+  "deck.settings-modal.review-pacing.preset-line": "Using the {preset} preset",
+  "deck.settings-modal.review-pacing.preset-description": "Presets set this deck's baseline pace. Anything you change below applies to this deck only.",
   "deck.settings-modal.review-pacing.default-preset-label": "Default",
+  "deck.settings-modal.review-pacing.general-heading": "General",
+  "deck.settings-modal.review-pacing.general-description": "How cards are presented while you study.",
+  "deck.settings-modal.review-pacing.limits-heading": "Daily Limits",
+  "deck.settings-modal.review-pacing.limits-description": "How much this deck puts in front of you each day.",
+  "deck.settings-modal.review-pacing.limits-hint": "Set either to 0 for no limit.",
   "deck.settings-modal.review-pacing.advanced-toggle": "Advanced",
-  "deck.settings-modal.review-pacing.advanced.scheduling-heading": "Scheduling Algorithm",
-  "deck.settings-modal.review-pacing.advanced.scheduling-description": "Fine-tune how the spaced repetition scheduler paces your reviews.",
+  "deck.settings-modal.review-pacing.advanced-back": "Back",
+  "deck.settings-modal.review-pacing.scheduling-heading": "Scheduling",
+  "deck.settings-modal.review-pacing.scheduling-description": "How the scheduler spaces your reviews over time.",
   "deck.settings-modal.review-pacing.desired-retention-label": "Desired Retention",
   "deck.settings-modal.review-pacing.desired-retention-tooltip": "Higher values mean more frequent reviews, but better recall.",
   "deck.settings-modal.review-pacing.max-interval-label": "Maximum Interval",
